Split build graph dependency tests
